requirements.txt
langchain
langgraph
tavily-python
polygon
langchain_community
langchain_google_genai
streamlit
fastapi[all]
uvicorn
langchain-pinecone
pypdf
python-box
ensure
google-generativeai
python-dotenv
langchain_huggingface
pyYAML
pinecone
# -e .